css_parse: Add SourceCursor, SourceCursorSink, impl SourceCursorSink for fmt::Write (#312)
We have the CusrorFmtSink which acts as a Sink to take in Cursors and
write them out to the given fmt::Write object, but
this has a fair bit of incidental complexity:

- CusrorFmtSink needs to be given the Source Text, so it can pull out
idents and such.
- It needs to retain and have logic around the last token so that
contiguous tokens which create ambiguous single tokens
	can be separated with whitespace.
- Because of this, it can only accept cursors that match the original
source text, not new ones (say from a transform).
- This makes it tricky to compose.

These complexities mean it's not very simple, and while it has its
purpose, it's kind of the lowest level Sink we can
have. These issues could in theory be resolved if we change how cursors
are represented. Cursors are cheap because they are a
u64, and avoiding having to store a `&str` which inflates the size and
also has to carry with it a lifetime. Adding
this to a Cursor would increase complexity & size of AST nodes. But it
means we lose the source text.

If we _reintroduce_ the source text to a Cursor, we might get a struct
like SourceCursor - a struct which holds both the
cursor and a string slice of the original source text. These
SourceCursors can be used in places a little more
transitory, where the struct size isn't so important and what _is_
important it getting to the underlying string data,
say, for example, when writing cursors to a fmt::Write.

With the new SourceCursor struct, we can create a
CursorToSourceCursorSink which takes Cursors, and a given source text,
and smooshes them into a set of SourceCursors. The SourceCursors can
then be used in interesting ways, such as fed
directly into a fmt::Write... but we need some kind of trait to do that.

So we _also_ can implement a SourceCursorSink trait. This is just like
CursorSink but rather than append taking a Cursor, it takes a
SourceCursor. With this in place, we can now introduce a blanket impl
for SourceCursorSink for fmt::Write, and the picture becomes more
complete.

So this change introduces those things: SourceCursor - a "fat" cursor
with source_text. SourceCursorSink, a CursorSink that appends
SourceCursors, and impls SourceCursorSink for fmt::Write, negating much
of the need for CursorFmtSink!