Simplify Function<'a, FT, T> to Function<FT, T> (#320)
The Function dynamic type was a little hastily built, and as a
consequence the struct definition ended up requiring all
of the common traits, plus the `'a` lifetime used for the bump
allocator. It didn't need this (evident by the fact is
was stuffed into PhantomData).

This change drops the trait requirements (requiring them only when
needed) and as a consequence drops the `'a` lifetime.
This also has a rather large fanout effect because there aren't many
functions that actually _needed_ the lifetime, but
the assumption was baked in all the way to generate.rs...

So this change not only drops the lifetime from all structs, but also
tweaks generate.rs to only supply the lifetime to
the functions that need it (via the usual routes).