css_lexer: Ensure Span::DUMMY can be used without being added to (#341)
Many types can end up having no Span (e.g. an Option), but using
Span::ZERO meant that their start character was always
0, and when added to other spans, it would remain `0`. For example a
`(Option<T![Ident]>, T![Ident])` was guaranteed to
have a span of `0` which is not correct.

This change alters the desired sentinel type for these empty spans to be
DUMMY. This resolves the issue of spans
_starting_ at `0` but means spans _end_ at `u32::MAX` which is equally
undesirable. Consequently this also changes `add()` to ensure that if
`Span::DUMMY` is passed as one of the operands, it'll simply return the
other operand, avoiding ever adding a DUMMY to another span.

This should resolve many Span issues across the codebase.