css_ast: Add assert_visits! test & some tests (#345)
Our visit functionality is not suitably tested, which leaves opportunity
for holes everywhere.

This changes adds an `assert_visits!` macro which, given a string to
parse, and a type to parse against, will visit all
possible visit methods and accumulate the visited types, collecting them
into a list which can be compared against the
given arguments. The result of this assertion is an accurate ordered
list of all the types visited during a traversal of
the AST.