csskit_proc_macro: Only add #[visit] attrs if derive(Visitable), avoid Parse impl if derive(Parse) (#357)
In order to use #[syntax] in more places it needs to be a little more
flexible about when it adds #[visit] attributes,
only adding them when something actually implements Visitable. We can't
tell during a proc macro if a type is going to
impl Visitable _manually_, but we do have a good chance of finding out
if it derives(Visitable), by reading the derive
attribute.

So this change checks if a type has derive(Visitable), and only adds
`#[visit(skip)]` _if_ that derive is present.

This also makes sense to check if derive(Parse) is present, because if
so we can simply skip implementing Parse in those
instances. So we also do that here.