csskit_derives: Add support for parsing as keyword (#361)
csskit_derives would always parse an Ident as an Ident. This is
different to generate.rs which would use a Keywords
struct/enum.

This change adds a new #[parse(keyword = ...)] syntax which allows the
struct to lean on that type to parse & get an
Ident back, which negates the need to pack that keyword into the struct.