css_lexer: Stop storing length in flags. (#392)
There really is little point to storing the `len_utf8()` of a char in
the Tokens flag space, as this data is already
encoded into the char itself.

This change removes the `Kind::Delim` flag commentary, and makes those
flags reserved, which simplifies the construction
of Delim / single char tokens. The len() function can then simply
delegate to char().unwrap().len_utf8().