css_lexer/css_parse: Introduce AssociatedWhitespaceRules (#393)
Rules around whitespace in CSS are a little complex, as they're often
contextual. In
https://github.com/csskit/csskit/pull/388 we spotted a bug where
`calc(100vh - 68px)` would be potentially re-serialized as
``calc(100vh-68px)` which is invalid, and while that fix was proper, it
didn't tell a complete story of some of the complexities around
whitespace, especially with regard to re-serializing in different modes
(such as a "prettier" version vs a "compact" version.

This change introduces more context embedded into the Tokens themselves,
as an AssociatedWhitespaceRules set of bitflags. These bitflags
determine the rules for when whitespace must be enforced either side of
a delimiter, which can be more strictly enforced inside of a set of
`ComponentValues`, or more loosely applied in other places. It also
allows for us to ban whitespace for delim-ident adjacent tokens such as
`:hover`, `::before`, `$var` which would break if a "prettier" formatter
were to try and separate with whitespace.

Overall this provides a small but useful change which we can leverage
more in css_ast as the need arises, but this in of
itself also should alleviate some issues like the bugs spotted in
https://github.com/csskit/csskit/pull/388 in a more
robust but also more flexible manner.