move SourceCursor into css_lexer & refine (#398)
SourceCursor is quite useful but one issue is that the span data becomes
inconsistent with overlays, due to the way a SourceCursor reads into its
string data. So this change alters SourceCursor so that the span data is
not related to the string slice, and instead the string slice holds the
entirety of the character data for that token (as opposed to being a
superset).

This change also meant that we could decouple Cursor from its ability to
write out data, and instead push more of that responsibility into
SourceCursor. This _also_ means getting rid of SpanContents which was an
awkward type mostly just used to get the line & column data.
SourceCursor replaces SpanContents's intent beyond that.

This then _also_ makes SourceCursor better placed in css_lexer instead
of css_parse, so that also happens in this
change.

Lastly, all downstream APIs need updating. Mostly small API changes, but
also making some functions const.