css_ast: Shrink color function size by using CommaOrSlash struct (#407)
RgbFunctions & similar can be represented in the "legacy comma" style,
or the style without commas and a `/` for alpha.
The struct for these allowed both the comma and slash to appear
simultaneously, but this would be a parse error in browsers. We can
reasonably enforce the same level of strictness, which would shrink the
struct size for these types, which cascades quite nicely all the way up
to StyleValue & Rule, so a worthwhile win.