Re-introduce string interning across codebase
This rather substantial change re-introduces a form of string interning back into css_ast. We used to have interned strings using Servo's string_cache crate, which used PHF map under the hood, the problem with the particular implementation was that the resulting atom!() macros were a little awkward to use, but importantly the Lexer couldn't pack atom data into the token because A) it was too large and IV) it would couple the Lexer to CSS atoms.

This change introduces a new derive_atom_set crate. This crate allows us to generate atom sets that we can then try to pack into Lexer Token data. The bits from the Token can be extracted and reconsituted into an Atom which css_ast can leverage, instead of trying to construct PHF maps everywhere.

This change also significantly cleans up the css_ast codebase by leveraging this pattern, which also allows us to drop a whole bunch of structs/macros/generics whose purpose was mostly to alleviate the labor of not having string interning.

This is a giant commit and I'd prefer it not to be - I'd like to split it up and make it more modular, but it's already over a week of work and I don't really want to tidy it up any more, so I'm just dropping it wholesale. YOLO