css_parse: add trivia output (#499)
Trivia tokens have so far been swallowed up by the parser, never to be
emitted back out as cursors. This change attempts
to address this by making sure the ParserReturn properly emits trivia
cursors when `with_trivia()` is enabled.

This needed some architectural changes, as trivia was simply collected
into a Vec without association, but this causes
issues for example a function like `foo(/*bar*/ bar, baz)` might be
re-tokenised as `foo(/*bar*/ baz, bar)` which is
likely wrong. Instead we introduce an associated cursor with each vec of
trivia, this way trivia such as comments are
attached to the trailing token, meaning in the above example the
`/*bar*/` comment and whitespace token after that (` `)
are both attached to the `bar` token.

Sadly, this has the unfortunate side effect that the whitespace is also
attached, which means if a node restructures the
cursors, we might end up with for e.g. `foo(bar, baz)` becoming `foo(
baz, bar)` (note the additional whitespace after
`foo(`). This is because the whitespace in the token stream `, baz`.
This can be resoled two ways: simply avoiding the
CursorWriteSink and always using it in combination with the
Compact/Format write sink which handles whitespace more
sensibly, but also eventually perhaps writing a OrderedCursorSink which
buffers cursors so it is able to emit them in
source order (which is likely necessary to handle IDE code actions
without incidentally formatting code).