css_parse: Avoid having to manually flush Cursor Sinks by emitting Eof. (#501)
We had a bug whereby trailing Trivia was not being correctly emitted due
to the need to flush various CursorSinks. The
problem is the sinks can be combined in arbitrary orders and may need to
be flushed in different orders. Rather than
this being the responsibility of callsites, we can instead emit an Eof
token in ToCursors which can act as a sentinel
value for the sinks to flush.

So this change implements that: ParserReturn ToCursors emits an Eof, and
both CursorOrderedSink and
CursorInterleaveSink will flush their respective buffers based on this
signal. CursorOverlaySink needs some small
changes to ensure that it doesn't emit Eof from Overlays, as an Eof
effectively closes the sink and should only be
emitted once. Consequently some debug asserts have been added to ensure
sinks only see one Eof.