chromashift: Add ability to extract alpha, TryInto named color (#567)
Working with chromashift a little and I noticed two issues which I
wanted to improve: it can be a little clunky to get
an alpha channel from a `Color` as you first have to either match which
colour you're referring to or convert it to a
known color (e.g. srgb) to then get the alpha. It would be much easier
if there were a unilateral trait to simply get
the alpha from any colour or indeed the generic `Color` enum.
Additionally I wanted to convert a colour to a named colour,
and this proved awkward to do as I would have to re-introduce the data
from Named, or iterate and match (slow).

So this commit introduces those two improvements:

- The ToAlpha trait adds a `to_alpha(&self) -> f32` method to all
colours, making it trivial to just get the alpha
channel of the colour. This also adds convenience methods of
`fully_opaque` and `fully_transparent`.
- Named now implements TryFrom<?>, where `?` is any of the colors,
returning `Result<Named>`. If the colour was not
fully opaque then it'll return `ToNamedError::NotOpaque`. If a named
colour wasn't found then it'll return
	`ToNamedError::NoMatchingName`.