csskit_proc_macro: drop hardcoded types, rely on ast imports more (#589)
The csskit_proc_macro Def & Generate code had long match statements
matching a bunch of hard coded types, mostly just to
provide range checks, lifetime matching, and aliasing, but this is a
little redundant.

This change drops all of those, and instead expects the types from the
crate (i.e. css_ast) to match the names in the
grammar.