css_value_definition_parser: support quoted literal puncts, fixed groups (#611)
Testing various syntax combinations I noticed that quoted puncts were
failing because they're parsed as
Lit::Char/Lit::Str which wasn't caught.

This change now recognises those and produces valid Defs for them.

In addition using a group syntax with a fixed range failed, so while
`<length>{5}` worked, `[<length>]{5}` did not. This
was because of an invariant assumption that groups must always have a
range between two values, but it's valid for them
to be fixed also.

So this change allows for such fixed ranges.