css_ast/csskit_proc_macro: Support `,` literals better. (#638)
https://github.com/w3c/csswg-drafts/pull/13018 changed the syntax of
Cursor from using `#?` to `,]*`. This subtle change
means trailing commas are always required. We didn't support this for a
couple of reasons:

- Punct wasn't being rendered correctly due to quoting of the char, so
`T![,]` was rendering as `T![',']` which isn't a
known macro. css_parse could have included quoted variants of all
characters, or we could have updated
	csskit_proc_macro to unquote the chars, which is what this change does.

- `T![,]` wasn't Visitable, meaning the `Vec<'a, (X, T![,])>` wasn't
Visitable either. Both `Vec` & `(A, B)` _are_
visitable, so it's only `T![,]` that needs to be to make the whole thing
work. This change makes `T![,]` Visitable.