css_ast/css_parse: Properly handle parse failures for unknown Rules vs Declarations (#700)
We've had a pretty gnarly bug for a long time because of the way
properties vs
nested rules are parsed: if a property is attempted to be parsed but
fails, we
simply try to parse it as a rule instead. If a Rule fails, we parse a
bad
declaration because the spec says to "consume the remnants of a bad
declaration". This causes myriad issues:

1. Rules and Declarations can sometimes be grammatically ambiguous,
which means
we sometimes swallow entire rules up as declarations, or entire
declarations
	 as rules.
2. This has a knock-on effect of implicitly re-ordering rules and
declarations,
as right now there is no mechanism of interleaving declarations. This
can do
bad things during minification if a declaration accidentally ends up at
the
bottom of the rules or a rule is appended to the declaration list and
ends
	 up above other rules.

The solution to this is kind of complex, and it's made more complex by
us being
a generic parser rather than a specific engine: engines don't need to
account
for "unknown declarations" or "unknown rules", they're just parse errors
with
well defined recovery, but we need to account for them and treat them
specially.

This means the solution boils down to roughly:

1. Catalog all known property IDs and use that as a heuristic to
determine if a
declaration failed to parse because we don't recognise it (even if it's
	 known) or because it's actually not known by any engine or spec.
2. Improve heuristics for detecting invalid declarations. The CSS spec
has notes
   for a few ideas here, including that a dashed ident will always be a
   declaration and never the start of a rule.
3. Don't just rely on parse failures to rewind & reparse
declarations/rules;
instead check if it was a successful but _unknown_ declaration, and try
parsing as a rule instead, and if that is _also_ unknown then finally
parse
as a declaration (this is the "consume the remnants of a bad
declaration"
	 piece).

These changes combined put us in a much better position where we're
parsing more
properties & nested rules more successfully.

However there's still more work to do: we could still improve peeking to
detect
these earlier, but also look at other strategies like forking parsers.
We also
need to provide the ability for interleaving declarations and rules to
avoid
accidentally re-ordering rules/declarations in other ways.