css_parse: Consume whitespace as trivia around non-descendant combinators (#723)
A selector like `a > b` is comprised of three parts: a tag selector, a
child
combinator, and another tag selector. The whitespace around the `>`
should be
treated as whitespace, but at current it is treated as a descendant
combinator.

This change improves parsing of the Selector trait so that it consumes
whitespace if it can peek ahead to an adjacent combinator, and consumes
whitespace _after_ parsing a combinator. Adjacent combinators are
disallowed so
this change improves correctness of selector parsing.