csskit_ast: Rewrite QuerySelector to use css_parse (#724)
QuerySelector was initially written as a rough prototype manual
implementation
of selectors to get the matcher working, and work around the issues of
adjacent
combinators which is now fixed in #723.

This change rewrites all of QuerySelector to know use the css_parse
Selector
crate, which allows for a much simpler implementation. This also fixes
up
find.rs to use the more traditional `parse_entirely` now it's using
conventional traits.