csskit_highlight/csskit/chromashift: Support owo-colors (#766)
While we've traditionally used anstyle for ascii styling, miette
supports owo-colors, and to add syntax highlighting it
requires data returned with the owo-colors `Style` struct. Both
libraries are
fairly similar and we don't have a particular allegiance to one, but
given
miette has a requirement, we need to support owo-colors.

This change implements `owo-colors` as an optional feature flag for
chromashift,
which enables turning `chromashift::Srgb` into `owo-colors::Rgb`. The
same flag
in csskit_highlight enables `get_owo_style()` on `DefaultAnsiTheme`
which the
`AnsiHighlightCursorStream` uses if the flag is enabled. Lastly, csskit
gets the
same flag and will switch from using anstyle to owo-colors. We'll likely
keep
both around for a while just in case, but eventually drop anstyle if
owo-colors
works out fine.